mock-merchant/requirements.txt
# Mock Merchant Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Templates & Static Files
jinja2==3.1.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# Cryptography (for TAP verification)
cryptography==42.0.0

# Utilities
python-dotenv==1.0.0
